feat: extracted list
