+++ src/renderer/features/fellowship-activity-feed/components/ActivityList.tsx
@@ -1,11 +1,16 @@
+import { useUnit } from 'effector-react';
 import { memo } from 'react';
 
-import { ActivityListBase } from './ActivityListBase';
+import { activityFeed } from '../model/list';
+
+import { ActivityListView } from './ActivityListView';
 
 export const ActivityList = memo(() => {
+  const feed = useUnit(activityFeed.$activityFeed);
+
   return (
     <div className="flex flex-col gap-3 py-4 pb-3">
-      <ActivityListBase limit={20} />
+      <ActivityListView limit={20} feed={feed} />
     </div>
   );
 });
