+++ src/renderer/features/fellowship-activity-feed/components/ActivityModal.tsx
@@ -1,17 +1,78 @@
 import { useUnit } from 'effector-react';
-import { type PropsWithChildren } from 'react';
+import { orderBy } from 'lodash';
+import { type PropsWithChildren, useDeferredValue, useMemo, useState } from 'react';
 
 import { useI18n } from '@/shared/i18n';
-import { nullable } from '@/shared/lib/utils';
-import { Modal } from '@/shared/ui-kit';
+import { nullable, performSearch, truncate } from '@/shared/lib/utils';
+import { Button, EmptyList } from '@/shared/ui';
+import { Modal, SearchInput, Select } from '@/shared/ui-kit';
+import { identityService } from '@/domains/network';
 import { fellowshipActivityFeedFeature } from '../model/feature';
+import { identityModel } from '../model/identity';
+import { activityFeed } from '../model/list';
 
-import { ActivityListBase } from './ActivityListBase';
+import { ActivityListView } from './ActivityListView';
+import { getDescription } from './utils';
+
+type OrderKey = 'duration-asc' | 'duration-desc' | 'name-asc' | 'name-desc';
+
+const orderVariants: Record<OrderKey, { field: string; direction: 'asc' | 'desc' }> = {
+  'duration-asc': { field: 'duration', direction: 'asc' },
+  'duration-desc': { field: 'duration', direction: 'desc' },
+  'name-asc': { field: 'name', direction: 'asc' },
+  'name-desc': { field: 'name', direction: 'desc' },
+};
 
 export const ActivityModal = ({ children }: PropsWithChildren) => {
   const { t } = useI18n();
 
   const input = useUnit(fellowshipActivityFeedFeature.input);
+  const feed = useUnit(activityFeed.$activityFeed);
+  const identities = useUnit(identityModel.$list);
+
+  const [query, setQuery] = useState('');
+  const [orderKey, setOrderKey] = useState<OrderKey>('duration-asc');
+
+  const deferredQuery = useDeferredValue(query);
+
+  const clearSearch = () => setQuery('');
+
+  const records = useMemo(
+    () =>
+      feed.map(record => {
+        const identity = identities[record.accountId];
+        return {
+          ...record,
+          name: identity ? identityService.getFullName(identity) : undefined,
+          description: getDescription(record, t),
+        };
+      }),
+    [identities, feed, t],
+  );
+
+  const filteredList = useMemo(() => {
+    return performSearch({
+      query: deferredQuery,
+      queryMinLength: 2,
+      records,
+      weights: {
+        name: 1,
+        description: 0.75,
+        address: 0.5,
+        type: 0.5,
+        wish: 0.5,
+        accountId: 0.5,
+      },
+    });
+  }, [deferredQuery, records]);
+
+  const isNothingFound = !!records.length && !filteredList.length;
+  const orderVariant = orderKey ? orderVariants[orderKey] : null;
+
+  const sortedList = useMemo(
+    () => (orderVariant ? orderBy(filteredList, orderVariant.field, orderVariant.direction) : filteredList),
+    [filteredList, orderVariant],
+  );
 
   if (nullable(input)) return children;
 
@@ -19,11 +80,49 @@ export const ActivityModal = ({ children }: PropsWithChildren) => {
     <Modal size="md" height="lg">
       <Modal.Trigger>{children}</Modal.Trigger>
       <Modal.Title close>{t('fellowship.activityFeed.activityModal.title')}</Modal.Title>
+      <Modal.HeaderContent>
+        <div className="flex gap-x-2 bg-main-app-background px-5 py-4">
+          <div className="inline grow">
+            <SearchInput
+              placeholder={t('fellowship.activityFeed.activityModal.search-placeholder')}
+              value={query}
+              onChange={setQuery}
+            />
+          </div>
+
+          <div className="w-[150px]">
+            <Select
+              placeholder={t('fellowship.activityFeed.activityModal.sort.placeholder')}
+              value={orderKey}
+              onChange={setOrderKey}
+            >
+              <Select.Item value="duration-asc">
+                {t('fellowship.activityFeed.activityModal.sort.duration-asc')}
+              </Select.Item>
+              <Select.Item value="duration-desc">
+                {t('fellowship.activityFeed.activityModal.sort.duration-desc')}
+              </Select.Item>
+              <Select.Item value="name-asc">{t('fellowship.activityFeed.activityModal.sort.name-asc')}</Select.Item>
+              <Select.Item value="name-desc">{t('fellowship.activityFeed.activityModal.sort.name-desc')}</Select.Item>
+            </Select>
+          </div>
+        </div>
+      </Modal.HeaderContent>
       <Modal.Content>
         <div className="bg-main-app-background">
-          <ActivityListBase limit={Number.POSITIVE_INFINITY} isFullVersion={true}>
-            {children}
-          </ActivityListBase>
+          {isNothingFound && (
+            <EmptyList
+              title={t('fellowship.activityFeed.activityModal.nothing-found.title')}
+              message={t('fellowship.activityFeed.activityModal.nothing-found.description', {
+                query: truncate(query, 6, 6),
+              })}
+            >
+              <Button pallet="primary" variant="text" onClick={clearSearch}>
+                {t('fellowship.activityFeed.activityModal.nothing-found.clear')}
+              </Button>
+            </EmptyList>
+          )}
+          <ActivityListView limit={Number.POSITIVE_INFINITY} feed={sortedList} withFullAccountInfo />
         </div>
       </Modal.Content>
     </Modal>
